[fix]: [PL-23042]: Develop merge of Hotfix: Skip Delete permission in Hashicorp vault for Create/Update/Read operations (#30866)

+++ 890-sm-core/src/main/java/io/harness/encryptors/clients/HashicorpVaultEncryptor.java
@@ -63,7 +63,7 @@ public class HashicorpVaultEncryptor implements VaultEncryptor {
     while (true) {
       try {
         return HTimeLimiter.callInterruptible21(timeLimiter, Duration.ofSeconds(15),
-            () -> upsertSecretInternal(name, plaintext, accountId, null, vaultConfig));
+            () -> upsertSecretInternal(name, plaintext, accountId, null, vaultConfig, false));
       } catch (Exception e) {
         failedAttempts++;
         log.warn("encryption failed. trial num: {}", failedAttempts, e);
@@ -88,7 +88,7 @@ public class HashicorpVaultEncryptor implements VaultEncryptor {
     while (true) {
       try {
         return HTimeLimiter.callInterruptible21(timeLimiter, Duration.ofSeconds(5),
-            () -> upsertSecretInternal(name, plaintext, accountId, existingRecord, vaultConfig));
+            () -> upsertSecretInternal(name, plaintext, accountId, existingRecord, vaultConfig, false));
       } catch (Exception e) {
         failedAttempts++;
         log.warn("encryption failed. trial num: {}", failedAttempts, e);
@@ -142,12 +142,12 @@ public class HashicorpVaultEncryptor implements VaultEncryptor {
   }
 
   private EncryptedRecord upsertSecretInternal(String keyUrl, String value, String accountId,
-      EncryptedRecord existingRecord, VaultConfig vaultConfig) throws IOException {
+      EncryptedRecord existingRecord, VaultConfig vaultConfig, boolean deleteRequired) throws IOException {
     log.info("Saving secret {} into Vault {}", keyUrl, vaultConfig.getBasePath());
 
     // With existing encrypted value. Need to delete it first and rewrite with new value.
     String fullPath = getFullPath(vaultConfig.getBasePath(), keyUrl);
-    deleteSecret(accountId, EncryptedRecordData.builder().encryptionKey(keyUrl).build(), vaultConfig);
+
     String vaultToken = getToken(vaultConfig);
     boolean isSuccessful = VaultRestClientFactory.create(vaultConfig)
                                .writeSecret(String.valueOf(vaultToken), vaultConfig.getNamespace(),
@@ -158,7 +158,9 @@ public class HashicorpVaultEncryptor implements VaultEncryptor {
       if (existingRecord != null) {
         String oldFullPath = getFullPath(vaultConfig.getBasePath(), existingRecord.getEncryptionKey());
         if (!oldFullPath.equals(fullPath)) {
-          deleteSecret(accountId, existingRecord, vaultConfig);
+          if (deleteRequired) {
+            deleteSecret(accountId, existingRecord, vaultConfig);
+          }
         }
       }
       return EncryptedRecordData.builder().encryptionKey(keyUrl).encryptedValue(keyUrl.toCharArray()).build();
@@ -172,7 +174,7 @@ public class HashicorpVaultEncryptor implements VaultEncryptor {
   private EncryptedRecord renameSecretInternal(
       String keyUrl, String accountId, EncryptedRecord existingRecord, VaultConfig vaultConfig) throws IOException {
     char[] value = fetchSecretInternal(existingRecord, vaultConfig);
-    return upsertSecretInternal(keyUrl, new String(value), accountId, existingRecord, vaultConfig);
+    return upsertSecretInternal(keyUrl, new String(value), accountId, existingRecord, vaultConfig, true);
   }
 
   @Override
